requirements.txt
absl-py==2.3.1
accelerate==1.8.1
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autogluon==1.3.1
autogluon.common==1.3.1
autogluon.core==1.3.1
autogluon.features==1.3.1
autogluon.multimodal==1.3.1
autogluon.tabular==1.3.1
autogluon.timeseries==1.3.1
beartype==0.21.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
blis==1.3.0
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.7
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
coreforecast==0.0.15
cut-cross-entropy==25.1.1
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
defusedxml==0.7.1
diffusers==0.34.0
dill==0.3.8
distlib==0.3.9
docstring_parser==0.16
efficientnet_pytorch==0.7.1
einops==0.8.1
evaluate==0.4.4
fastai==2.8.2
fastcore==1.8.5
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fs==2.4.16
fsspec==2025.3.0
fugue==0.9.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
gluonts==0.16.2
google-api-core==2.25.1
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
hf_transfer==0.1.9
huggingface-hub==0.33.2
hyperopt==0.2.7
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mlforecast==0.13.6
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
namex==0.1.0
narwhals==1.46.0
networkx==3.4.2
nlpaug==1.1.11
nltk==3.8.1
numba==0.61.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.11
opt_einsum==3.4.0
optree==0.16.0
optuna==4.4.0
ordered-set==4.1.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pdf2image==1.17.0
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
plum-dispatch==2.5.7
preshed==3.0.10
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
py-spy==0.4.0
py4j==0.10.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smart_open==7.3.0.post1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
statsforecast==2.0.1
statsmodels==0.14.5
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.3
tokenizers==0.21.2
tomli==2.2.1
toolz==0.12.1
torch==2.6.0
torchmetrics==1.7.4
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
triad==0.9.8
triton==3.2.0
trl==0.19.1
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tyro==0.9.26
tzdata==2025.2
unsloth==2025.6.12
unsloth_zoo==2025.6.8
urllib3==2.5.0
utilsforecast==0.2.10
virtualenv==20.31.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
window_ops==0.0.15
wrapt==1.17.2
xformers==0.0.30
xgboost==3.0.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
